migrate `ButtonPrimaryInvert`
