+++ packages/app/src/canvas/ManageNominations/index.tsx
@@ -23,8 +23,7 @@ import { useBondedPools } from 'contexts/Pools/BondedPools';
 import { RevertPrompt } from './Prompts/RevertPrompt';
 import { CanvasSubmitTxFooter, CanvasFullScreenWrapper } from '../Wrappers';
 import { NotificationsController } from 'controllers/Notifications';
-import { ButtonHelp, ButtonPrimary } from 'ui-buttons';
-import { ButtonPrimaryInvert } from 'kits/Buttons/ButtonPrimaryInvert';
+import { ButtonHelp, ButtonPrimary, ButtonPrimaryInvert } from 'ui-buttons';
 
 export const ManageNominations = () => {
   const { t } = useTranslation('library');
