+++ packages/app/src/library/GenerateNominations/index.tsx
@@ -30,8 +30,7 @@ import { usePrompt } from 'contexts/Prompt';
 import { useFetchMehods } from './useFetchMethods';
 import type { AddNominationsType, GenerateNominationsProps } from './types';
 import type { AnyJson, AnyFunction } from '@w3ux/types';
-import { ButtonPrimaryInvert } from 'kits/Buttons/ButtonPrimaryInvert';
-import { ButtonMonoInvert } from 'ui-buttons';
+import { ButtonPrimaryInvert, ButtonMonoInvert } from 'ui-buttons';
 
 export const GenerateNominations = ({
   setters = [],
