+++ packages/app/src/modals/Connect/Ledger.tsx
@@ -9,8 +9,7 @@ import { useHelp } from 'contexts/Help';
 import LedgerLogoSVG from '@w3ux/extension-assets/Ledger.svg?react';
 import { useOverlay } from 'kits/Overlay/Provider';
 import { useNetwork } from 'contexts/Network';
-import { ButtonHelp } from 'ui-buttons';
-import { ButtonPrimaryInvert } from 'kits/Buttons/ButtonPrimaryInvert';
+import { ButtonHelp, ButtonPrimaryInvert } from 'ui-buttons';
 import { ButtonText } from 'kits/Buttons/ButtonText';
 import { ModalConnectItem } from 'kits/Overlay/structure/ModalConnectItem';
 import { ModalHardwareItem } from 'kits/Overlay/structure/ModalHardwareItem';
