+++ packages/app/src/pages/Nominate/NominationGeo/NominationGeoList/index.tsx
@@ -7,10 +7,9 @@ import { MotionContainer } from 'library/List/MotionContainer';
 import type { NomninationGeoListProps } from '../types';
 import { useNetwork } from 'contexts/Network';
 import { Separator } from 'kits/Structure/Separator';
-import { ButtonPrimaryInvert } from 'kits/Buttons/ButtonPrimaryInvert';
+import { ButtonPrimaryInvert, ButtonHelp } from 'ui-buttons';
 import { useStaking } from 'contexts/Staking';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
-import { ButtonHelp } from 'ui-buttons';
 import { useHelp } from 'contexts/Help';
 import { Node } from './Node';
 import { CardHeaderWrapper } from 'library/Card/Wrappers';
