+++ packages/ui-buttons/src/ButtonPrimaryInvert/index.module.scss
@@ -1,0 +1,12 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.btnPrimaryInvert {
+  border: 1px solid var(--accent-color-stroke);
+  border-radius: var(--button-border-radius-large);
+  color: var(--accent-color-stroke);
+}
+.btnPrimaryInvert_secondaryColor {
+  border: 1px solid var(--accent-color-secondary);
+  color: var(--accent-color-secondary);
+}
