+++ packages/ui-buttons/src/index.tsx
@@ -1,8 +1,9 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export * from './ButtonPrimary';
 export * from './ButtonHelp';
 export * from './ButtonMono';
 export * from './ButtonMonoInvert';
 export * from './ButtonOption';
+export * from './ButtonPrimary';
+export * from './ButtonPrimaryInvert';
